💄 Use primary monitor instead of current
IMO current_monitor is a bit misleading as it uses the current monitor
the app is in, not the current monitor the user interacts with.

This is not optimal in a multi-monitor setup, as you often end
up with spyglass on the monitor that is not your primary. primary_monitor
on the other hand uses, you guessed it, the primary monitor, which I
see as more logical.

The optimal solution would be to get the monitor that the user is
currently interacting with, but this is not trivial.